src/recommender.py
# from langchain_classic.chains.combine_documents import create_stuff_documents_chain
# from langchain_classic.chains import create_retrieval_chain
from langchain_classic.chains import RetrievalQA
from langchain_groq import ChatGroq
from src.prompt_template import get_anime_prompt


class AnimeRecommender:

    def __init__(self,retriever,api_key:str,model_name:str):

        self.llm = ChatGroq(api_key = api_key, model = model_name , temperature = 0)
        self.prompt = get_anime_prompt()

        # question_answer_chain = create_stuff_documents_chain(
        #     self.llm,
        #     self.prompt
        # )

        # self.qa_chain = create_retrieval_chain(retriever,question_answer_chain)
        self.qa_chain = RetrievalQA.from_chain_type(
            llm = self.llm,
            chain_type = 'stuff',
            retriever = retriever,
            return_source_documents = True,
            chain_type_kwargs = {"prompt": self.prompt}
        )

    def get_recommendation(self,query:str):

        result = self.qa_chain({"query" : query})
        return result['result']