requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain_groq
langchain_huggingface
chromadb
streamlit
pandas
python-dotenv
sentence-transformers